event_processor/requirements.txt
aiohttp==3.9.1
pydantic==2.5.3
pydantic_settings==2.1.0
kafka-python==2.0.2
